+++ Changelog.md
@@ -28,6 +28,7 @@
 * fixed exception when creating a new DataFile with default text in a non-existent directory
 * fixed improper file names when attempting to create a DataFile with `.` in the path
 * Fixed numbers with decimal places using `,` instead of `.` on systems set to European standards
+* Saving and loading Type variables is much nicer with generic types
 
 todo update the docs for 1.2